Refactoring to move all smartcard functionality into scard.rs
