+++ lib/srv/desktop/rdp/rdpclient/src/rdpdr/scard.rs
@@ -13,27 +13,508 @@
 // limitations under the License.
 
 use crate::piv;
-use ironrdp_pdu::{other_err, PduResult};
-use ironrdp_rdpdr::pdu::efs::DeviceControlResponse;
-use ironrdp_rdpdr::pdu::esc::{ReadCacheCall, ScardContext, ScardHandle, WriteCacheCall};
+use ironrdp_pdu::utils::CharacterSet;
+use ironrdp_pdu::{custom_err, other_err, PduResult};
+use ironrdp_rdpdr::pdu::efs::{DeviceControlRequest, DeviceControlResponse, NtStatus};
+use ironrdp_rdpdr::pdu::esc::{
+    rpce, CardProtocol, CardState, CardStateFlags, ConnectCall, ConnectReturn, ContextCall,
+    EstablishContextReturn, GetDeviceTypeIdCall, GetDeviceTypeIdReturn, GetReaderIconReturn,
+    GetStatusChangeCall, GetStatusChangeReturn, HCardAndDispositionCall, ListReadersReturn,
+    LongReturn, ReadCacheCall, ReadCacheReturn, ReaderStateCommonCall, ReturnCode, ScardCall,
+    ScardContext, ScardHandle, ScardIoCtlCode, StatusReturn, TransmitCall, TransmitReturn,
+    WriteCacheCall,
+};
+use iso7816::Command as CardCommand;
 use std::collections::HashMap;
 use uuid::Uuid;
 
 #[derive(Debug)]
-pub struct Contexts {
+pub struct ScardBackend {
+    /// contexts holds all the active contexts for the server, established using
+    /// SCARD_IOCTL_ESTABLISHCONTEXT. Some IOCTLs are context-specific and pass it as argument.
+    ///
+    /// contexts also holds a cache and connected smartcard handles for each context.
+    contexts: Contexts,
+    uuid: Uuid,
+    cert_der: Vec<u8>,
+    key_der: Vec<u8>,
+    pin: String,
+}
+
+impl ScardBackend {
+    pub fn new(cert_der: Vec<u8>, key_der: Vec<u8>, pin: String) -> Self {
+        Self {
+            contexts: Contexts::new(),
+            uuid: Uuid::new_v4(),
+            cert_der,
+            key_der,
+            pin,
+        }
+    }
+
+    pub fn handle(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: ScardCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        match req.io_control_code {
+            ScardIoCtlCode::AccessStartedEvent => match call {
+                ScardCall::AccessStartedEventCall(_) => self.handle_access_started_event(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::EstablishContext => match call {
+                ScardCall::EstablishContextCall(_) => self.handle_establish_context(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::ListReadersW => match call {
+                ScardCall::ListReadersCall(_) => self.handle_list_readers(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::GetStatusChangeW => match call {
+                ScardCall::GetStatusChangeCall(call) => self.handle_get_status_change(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::ConnectW => match call {
+                ScardCall::ConnectCall(call) => self.handle_connect(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::BeginTransaction => match call {
+                ScardCall::HCardAndDispositionCall(_) => self.handle_begin_transaction(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::Transmit => match call {
+                ScardCall::TransmitCall(call) => self.handle_transmit(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::StatusW | ScardIoCtlCode::StatusA => match call {
+                ScardCall::StatusCall(_) => self.handle_status(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::ReleaseContext => match call {
+                ScardCall::ContextCall(call) => self.handle_release_context(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::EndTransaction => match call {
+                ScardCall::HCardAndDispositionCall(_) => self.handle_end_transaction(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::Disconnect => match call {
+                ScardCall::HCardAndDispositionCall(call) => self.handle_disconnect(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::Cancel => match call {
+                ScardCall::ContextCall(call) => self.handle_cancel(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::IsValidContext => match call {
+                ScardCall::ContextCall(_) => self.handle_is_valid_context(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::GetDeviceTypeId => match call {
+                ScardCall::GetDeviceTypeIdCall(call) => self.handle_get_device_type_id(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::ReadCacheW => match call {
+                ScardCall::ReadCacheCall(call) => self.handle_read_cache(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::WriteCacheW => match call {
+                ScardCall::WriteCacheCall(call) => self.handle_write_cache(req, call),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            ScardIoCtlCode::GetReaderIcon => match call {
+                ScardCall::GetReaderIconCall(_) => self.handle_get_reader_icon(req),
+                _ => Self::unsupported_combo_error(req.io_control_code, call),
+            },
+            _ => Err(custom_err!(
+                "TeleportRdpdrBackend::handle_scard_call",
+                SmartcardBackendError(format!(
+                    "received unhandled ScardIoCtlCode: {:?}",
+                    req.io_control_code
+                ))
+            )),
+        }
+    }
+
+    fn handle_access_started_event(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_establish_context(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let ctx = self.contexts.establish();
+
+        Self::dev_ctl_resp_return(req, EstablishContextReturn::new(ReturnCode::Success, ctx))
+    }
+
+    fn handle_list_readers(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Self::dev_ctl_resp_return(
+            req,
+            ListReadersReturn::new(ReturnCode::Success, vec![TELEPORT_READER_NAME.to_string()]),
+        )
+    }
+
+    fn handle_get_status_change(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: GetStatusChangeCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let timeout = call.timeout;
+        let context_id = call.context.value;
+
+        if timeout != TIMEOUT_INFINITE && timeout != TIMEOUT_IMMEDIATE {
+            // We've never seen one of these but we log a warning here in case we ever come
+            // across one and need to debug a related issue.
+            warn!(
+                "logic for a non-infinite/non-immediate timeout [{}] is not implemented",
+                timeout
+            );
+        }
+
+        let get_status_change_ret = Self::create_get_status_change_return(call);
+
+        // We have no status change to report, cache a response
+        // for later in case we get an SCARD_IOCTL_CANCEL.
+        if Self::has_no_change(&get_status_change_ret) {
+            if timeout != TIMEOUT_INFINITE {
+                return Err(other_err!(
+                    "TeleportRdpdrBackend::handle_list_readers",
+                    "got no change for non-infinite timeout",
+                ));
+            }
+
+            // Received a GetStatusChangeCall with an infinite timeout, so we're adding
+            // a corresponding DeviceControlResponse holding a GetStatusChangeReturn
+            // with its return code set to SCARD_E_CANCELLED to this Context. This value will
+            // be returned when we get an SCARD_IOCTL_CANCEL call for this Context.
+            self.contexts.set_scard_cancel_response(
+                context_id,
+                DeviceControlResponse::new(
+                    req,
+                    NtStatus::Success,
+                    Box::new(GetStatusChangeReturn::new(
+                        ReturnCode::Cancelled,
+                        get_status_change_ret.into_inner().reader_states,
+                    )),
+                ),
+            )?;
+
+            debug!("blocking GetStatusChange call indefinitely (since our status never changes) until we receive an SCARD_IOCTL_CANCEL");
+
+            return Ok(None);
+        }
+
+        // We have some status change to report, send it to the server.
+        Self::dev_ctl_resp_return(req, get_status_change_ret)
+    }
+
+    fn create_get_status_change_return(
+        call: GetStatusChangeCall,
+    ) -> rpce::Pdu<GetStatusChangeReturn> {
+        let mut reader_states = vec![];
+        for state in call.states {
+            match state.reader.as_str() {
+                // PnP is Plug-and-Play. This special reader "name" is used to monitor for
+                // new readers being plugged in.
+                "\\\\?PnP?\\Notification" => {
+                    reader_states.push(ReaderStateCommonCall {
+                        current_state: state.common.current_state,
+                        event_state: state.common.current_state,
+                        atr_length: state.common.atr_length,
+                        atr: state.common.atr,
+                    });
+                }
+                // This is our actual emulated smartcard reader. We always advertise its state as
+                // "present".
+                TELEPORT_READER_NAME => {
+                    let (atr_length, atr) = padded_atr::<36>();
+                    reader_states.push(ReaderStateCommonCall {
+                        current_state: state.common.current_state,
+                        event_state: CardStateFlags::SCARD_STATE_CHANGED
+                            | CardStateFlags::SCARD_STATE_PRESENT,
+                        atr_length,
+                        atr,
+                    });
+                }
+                // All other reader names are unknown and unexpected.
+                _ => {
+                    warn!(
+                        "got unexpected reader name [{}], ignoring",
+                        state.reader.as_str()
+                    );
+                    reader_states.push(ReaderStateCommonCall {
+                        current_state: state.common.current_state,
+                        event_state: CardStateFlags::SCARD_STATE_CHANGED
+                            | CardStateFlags::SCARD_STATE_UNKNOWN
+                            | CardStateFlags::SCARD_STATE_IGNORE,
+                        atr_length: state.common.atr_length,
+                        atr: state.common.atr,
+                    });
+                }
+            }
+        }
+
+        GetStatusChangeReturn::new(ReturnCode::Success, reader_states)
+    }
+
+    fn has_no_change(pdu: &rpce::Pdu<GetStatusChangeReturn>) -> bool {
+        pdu.into_inner_ref()
+            .reader_states
+            .iter()
+            .all(|state| state.current_state == state.event_state)
+    }
+
+    fn handle_connect(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: ConnectCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let handle = self.contexts.connect(
+            call.common.context,
+            call.common.context.value,
+            self.uuid,
+            &self.cert_der,
+            &self.key_der,
+            self.pin.clone(),
+        )?;
+
+        Self::dev_ctl_resp_return(
+            req,
+            ConnectReturn::new(ReturnCode::Success, handle, CardProtocol::SCARD_PROTOCOL_T1),
+        )
+    }
+
+    fn handle_begin_transaction(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_transmit(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: TransmitCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let cmd =
+            CardCommand::<TRANSMIT_DATA_LIMIT>::try_from(&call.send_buffer).map_err(|err| {
+                custom_err!(
+                    "TeleportRdpdrBackend::handle_transmit",
+                    SmartcardBackendError(format!(
+                        "failed to parse smartcard command {:?}: {:?}",
+                        &call.send_buffer, err
+                    ))
+                )
+            })?;
+
+        let card = self.contexts.get_card(&call.handle)?;
+        let resp = card.handle(cmd)?;
+
+        Self::dev_ctl_resp_return(
+            req,
+            TransmitReturn::new(ReturnCode::Success, None, resp.encode()),
+        )
+    }
+
+    fn handle_status(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let enc = match req.io_control_code {
+            ScardIoCtlCode::StatusW => CharacterSet::Unicode,
+            ScardIoCtlCode::StatusA => CharacterSet::Ansi,
+            _ => {
+                return Err(custom_err!(
+                    "TeleportRdpdrBackend::handle_status",
+                    SmartcardBackendError(format!(
+                        "got unexpected ScardIoCtlCode with a StatusCall: {:?}",
+                        req.io_control_code
+                    ))
+                ));
+            }
+        };
+
+        let (atr_length, atr) = padded_atr::<32>();
+
+        Self::dev_ctl_resp_return(
+            req,
+            StatusReturn::new(
+                ReturnCode::Success,
+                vec![TELEPORT_READER_NAME.to_string()],
+                // SPECIFICMODE state means that the card is ready to handle commands in a specific
+                // mode, no other negotiation is necessary. Real smartcards would probably negotiate
+                // some mode first.
+                CardState::SpecificMode,
+                CardProtocol::SCARD_PROTOCOL_T1,
+                atr,
+                atr_length,
+                enc,
+            ),
+        )
+    }
+
+    fn handle_release_context(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: ContextCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        self.contexts.release(call.context.value);
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_end_transaction(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_disconnect(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: HCardAndDispositionCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        self.contexts.disconnect(call.handle)?;
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_cancel(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: ContextCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let resp = self
+            .contexts
+            .take_scard_cancel_response(call.context.value)?;
+        if let Some(resp) = resp {
+            Ok(Some(resp))
+        } else {
+            // TODO: Currently we're just returning ReturnCode::Success here (based on awly's pre-
+            // IronRDP code). Should we instead be returning SCARD_E_CANCELLED (or something else)?
+            warn!("Received SCARD_IOCTL_CANCEL for a context without a pending SCARD_IOCTL_GETSTATUSCHANGEW");
+            Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+        }
+    }
+
+    fn handle_is_valid_context(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        // TODO: Currently we're always just sending ReturnCode::Success (based on awly's pre-
+        // IronRDP code). Should we instead be checking if we have such a context in our cache
+        // and returning an SCARD_E_INVALID_HANDLE (or something else)?
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_get_device_type_id(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: GetDeviceTypeIdCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        if self.contexts.exists(call.context.value) {
+            // Reader type describes the type of the physical connection to the smartcard reader (e.g.
+            // USB/serial/TPM). Type "vendor" means a proprietary vendor bus.
+            //
+            // See "ReaderType" in
+            // https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/smclib/ns-smclib-_scard_reader_capabilitiesconst SCARD_READER_TYPE_VENDOR: u32 = 0xF0;
+            const SCARD_READER_TYPE_VENDOR: u32 = 0xF0;
+            Self::dev_ctl_resp_return(
+                req,
+                GetDeviceTypeIdReturn::new(ReturnCode::Success, SCARD_READER_TYPE_VENDOR),
+            )
+        } else {
+            Err(custom_err!(
+                "TeleportRdpdrBackend::handle_get_device_type_id",
+                SmartcardBackendError(format!(
+                    "got GetDeviceTypeIdCall for unknown context [{}]",
+                    call.context.value
+                ))
+            ))
+        }
+    }
+
+    fn handle_read_cache(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: ReadCacheCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        let (data, return_code) = match self.contexts.read_cache(call)? {
+            Some(data) => (data, ReturnCode::Success),
+            None => (vec![], ReturnCode::CacheItemNotFound),
+        };
+        Self::dev_ctl_resp_return(req, ReadCacheReturn::new(return_code, data))
+    }
+
+    fn handle_write_cache(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        call: WriteCacheCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        self.contexts.write_cache(call)?;
+        Self::dev_ctl_resp_return(req, LongReturn::new(ReturnCode::Success))
+    }
+
+    fn handle_get_reader_icon(
+        &mut self,
+        req: DeviceControlRequest<ScardIoCtlCode>,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Self::dev_ctl_resp_return(
+            req,
+            GetReaderIconReturn::new(ReturnCode::UnsupportedFeature, vec![]),
+        )
+    }
+
+    /// This function returns the error for unsupported combinations of [`ScardIoCtlCode`] and [`ScardCall`].
+    fn unsupported_combo_error(
+        ioctl: ScardIoCtlCode,
+        call: ScardCall,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Err(custom_err!(
+            "TeleportRdpdrBackend::unsupported_combo_error",
+            SmartcardBackendError(format!(
+                "received unsupported combination of ScardIoCtlCode [{:?}] with ScardCall [{:?}]",
+                ioctl, call
+            ))
+        ))
+    }
+
+    /// Helper function for creating a DeviceControlResponse from a DeviceControlRequest and an
+    /// output buffer wrapped in an Ok(Some(...)), a common pattern in this object's methods.
+    fn dev_ctl_resp_return(
+        req: DeviceControlRequest<ScardIoCtlCode>,
+        output_buffer: impl rpce::Encode + 'static,
+    ) -> PduResult<Option<DeviceControlResponse>> {
+        Ok(Some(DeviceControlResponse::new(
+            req,
+            NtStatus::Success,
+            Box::new(output_buffer),
+        )))
+    }
+}
+
+#[derive(Debug)]
+struct Contexts {
     contexts: HashMap<u32, ContextInternal>,
     next_id: u32,
 }
 
 impl Contexts {
-    pub fn new() -> Self {
+    fn new() -> Self {
         Self {
             next_id: 1,
             contexts: HashMap::new(),
         }
     }
 
-    pub fn establish(&mut self) -> ScardContext {
+    fn establish(&mut self) -> ScardContext {
         let ctx_internal = ContextInternal::new();
         let id = self.next_id;
         self.next_id += 1;
@@ -42,7 +523,7 @@ impl Contexts {
         ctx
     }
 
-    pub fn connect(
+    fn connect(
         &mut self,
         ctx: ScardContext,
         id: u32,
@@ -56,47 +537,37 @@ impl Contexts {
         Ok(handle)
     }
 
-    pub fn disconnect(&mut self, handle: ScardHandle) -> PduResult<()> {
+    fn disconnect(&mut self, handle: ScardHandle) -> PduResult<()> {
         self.get_internal_mut(handle.context.value)?
             .disconnect(handle.value);
         Ok(())
     }
 
-    pub fn set_scard_cancel_response(
-        &mut self,
-        id: u32,
-        resp: DeviceControlResponse,
-    ) -> PduResult<()> {
+    fn set_scard_cancel_response(&mut self, id: u32, resp: DeviceControlResponse) -> PduResult<()> {
         self.get_internal_mut(id)?.set_scard_cancel_response(resp)
     }
 
-    pub fn take_scard_cancel_response(
-        &mut self,
-        id: u32,
-    ) -> PduResult<Option<DeviceControlResponse>> {
+    fn take_scard_cancel_response(&mut self, id: u32) -> PduResult<Option<DeviceControlResponse>> {
         Ok(self.get_internal_mut(id)?.take_scard_cancel_response())
     }
 
-    pub fn get_card(
-        &mut self,
-        handle: &ScardHandle,
-    ) -> PduResult<&mut piv::Card<TRANSMIT_DATA_LIMIT>> {
+    fn get_card(&mut self, handle: &ScardHandle) -> PduResult<&mut piv::Card<TRANSMIT_DATA_LIMIT>> {
         self.get_internal_mut(handle.context.value)?
             .get(handle.value)
             .ok_or_else(|| other_err!("Contexts::get_card", "unknown ScardHandle"))
     }
 
-    pub fn exists(&self, id: u32) -> bool {
+    fn exists(&self, id: u32) -> bool {
         self.contexts.contains_key(&id)
     }
 
-    pub fn read_cache(&mut self, call: ReadCacheCall) -> PduResult<Option<Vec<u8>>> {
+    fn read_cache(&mut self, call: ReadCacheCall) -> PduResult<Option<Vec<u8>>> {
         Ok(self
             .get_internal_mut(call.common.context.value)?
             .cache_read(&call.lookup_name))
     }
 
-    pub fn write_cache(&mut self, call: WriteCacheCall) -> PduResult<()> {
+    fn write_cache(&mut self, call: WriteCacheCall) -> PduResult<()> {
         self.get_internal_mut(call.common.context.value)?
             .cache_write(call.lookup_name, call.common.data);
         Ok(())
@@ -108,7 +579,7 @@ impl Contexts {
             .ok_or_else(|| other_err!("Contexts::get_internal_mut", "unknown context id"))
     }
 
-    pub fn release(&mut self, id: u32) {
+    fn release(&mut self, id: u32) {
         self.contexts.remove(&id);
     }
 }
@@ -200,7 +671,7 @@ const STATIC_ATR: [u8; 11] = [
     0x3B, 0x95, 0x13, 0x81, 0x01, 0x80, 0x73, 0xFF, 0x01, 0x00, 0x0B,
 ];
 
-pub fn padded_atr<const SIZE: usize>() -> (u32, [u8; SIZE]) {
+fn padded_atr<const SIZE: usize>() -> (u32, [u8; SIZE]) {
     let mut atr = [0; SIZE];
     let len = STATIC_ATR.len().min(SIZE);
     atr[..len].copy_from_slice(&STATIC_ATR[..len]);
@@ -208,8 +679,20 @@ pub fn padded_atr<const SIZE: usize>() -> (u32, [u8; SIZE]) {
 }
 
 pub const SCARD_DEVICE_ID: u32 = 1;
-pub const TELEPORT_READER_NAME: &str = "Teleport";
+const TELEPORT_READER_NAME: &str = "Teleport";
 // TRANSMIT_DATA_LIMIT is the maximum size of transmit request/response short data, in bytes.
-pub const TRANSMIT_DATA_LIMIT: usize = 1024;
-pub const TIMEOUT_INFINITE: u32 = 0xffffffff;
-pub const TIMEOUT_IMMEDIATE: u32 = 0;
+const TRANSMIT_DATA_LIMIT: usize = 1024;
+const TIMEOUT_INFINITE: u32 = 0xffffffff;
+const TIMEOUT_IMMEDIATE: u32 = 0;
+
+/// A generic error type for the TeleportRdpdrBackend that can contain any arbitrary error message.
+#[derive(Debug)]
+struct SmartcardBackendError(pub String);
+
+impl std::fmt::Display for SmartcardBackendError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:#?}", self)
+    }
+}
+
+impl std::error::Error for SmartcardBackendError {}
